backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-genai==0.3.0
firebase-admin==6.4.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
pytest==8.0.2
httpx==0.27.0
